backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
aiofiles 